+++ Source/CesiumRuntime/Public/CesiumRuntimeSettings.h
@@ -52,4 +52,18 @@ public:
    */
   UPROPERTY(Config, EditAnywhere, Category = "Experimental Feature Flags")
   bool EnableExperimentalOcclusionCullingFeature = false;
+
+  /**
+   * The number of requests to handle before each prune of old cached results
+   * from the database. Must restart Unreal for changes to take effect.
+   */
+  UPROPERTY(Config, EditAnywhere, Category = "Cache")
+  int RequestsPerCachePrune = 10000;
+
+  /**
+   * The maximum number of items that should be kept in the Sqlite database
+   * after pruning. Must restart Unreal for changes to take effect.
+   */
+  UPROPERTY(Config, EditAnywhere, Category = "Cache")
+  int MaxCacheItems = 4096;
 };